refactor: Move the provider's count method to the facade
