+++ core/context/src/main/java/com/milesight/beaveriot/context/api/DeviceServiceProvider.java
@@ -3,7 +3,6 @@ package com.milesight.beaveriot.context.api;
 import com.milesight.beaveriot.context.integration.model.Device;
 
 import java.util.List;
-import java.util.Map;
 
 /**
  * @author leon
@@ -20,8 +19,4 @@ public interface DeviceServiceProvider {
     Device findByIdentifier(String identifier, String integrationId);
 
     List<Device> findAll(String integrationId);
-
-    Map<String, Long> countByIntegrationIds(List<String> integrationIds);
-
-    Long countByIntegrationId(String integrationId);
 }
