+++ core/context/src/main/java/com/milesight/beaveriot/context/api/EntityServiceProvider.java
@@ -24,14 +24,6 @@ public interface EntityServiceProvider {
 
     void deleteByTargetId(String targetId);
 
-    long countAllEntitiesByIntegrationId(String integrationId);
-
-    Map<String, Long> countAllEntitiesByIntegrationIds(List<String> integrationIds);
-
-    long countIntegrationEntitiesByIntegrationId(String integrationId);
-
-    Map<String, Long> countIntegrationEntitiesByIntegrationIds(List<String> integrationIds);
-
     Entity findByKey(String entityKey);
 
     Map<String, Entity> findByKeys(List<String> entityKeys);
