+++ services/device/device-api/src/main/java/com/milesight/beaveriot/device/facade/IDeviceFacade.java
@@ -3,6 +3,7 @@ package com.milesight.beaveriot.device.facade;
 import com.milesight.beaveriot.device.dto.DeviceNameDTO;
 
 import java.util.List;
+import java.util.Map;
 
 public interface IDeviceFacade {
     List<DeviceNameDTO> fuzzySearchDeviceByName(String name);
@@ -14,4 +15,8 @@ public interface IDeviceFacade {
     List<DeviceNameDTO> getDeviceNameByKey(List<String> deviceKeys);
 
     DeviceNameDTO getDeviceNameByKey(String deviceKey);
+
+    Map<String, Long> countByIntegrationIds(List<String> integrationIds);
+
+    Long countByIntegrationId(String integrationId);
 }
