+++ services/device/device-service/src/main/java/com/milesight/beaveriot/device/service/DeviceServiceProviderImpl.java
@@ -14,7 +14,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
 
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -142,19 +141,6 @@ public class DeviceServiceProviderImpl implements DeviceServiceProvider {
                 .findAll(f -> f.eq(DevicePO.Fields.integration, integrationId))
                 .stream().map(deviceConverter::convertPO)
                 .toList();
-    }
-
-    @Override
-    public Map<String, Long> countByIntegrationIds(List<String> integrationIds) {
-        List<Object[]> res = deviceRepository.countByIntegrations(integrationIds);
-        Map<String, Long> result = new HashMap<>();
-        res.forEach((Object[] o) -> result.put((String) o[0], (Long) o[1]));
-        return result;
-    }
-
-    @Override
-    public Long countByIntegrationId(String integrationId) {
-        return deviceRepository.count(f -> f.eq(DevicePO.Fields.integration, integrationId));
     }
 
     private boolean deviceAdditionalDataEqual(Map<String, Object> arg1, Map<String, Object> arg2) {
