+++ services/user/user-service/pom.xml
@@ -96,6 +96,11 @@
         </dependency>
         <dependency>
             <groupId>com.milesight.beaveriot</groupId>
+            <artifactId>entity-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
+            <groupId>com.milesight.beaveriot</groupId>
             <artifactId>permission-api</artifactId>
             <version>${project.version}</version>
         </dependency>
